+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+  - Nuvlabox status - Update heartbeat even if NuvlaBox encounter a spec
+    validation issue
   - Nuvlabox-status - Use delete attributes
   - Job - Use delete-attributes
   - Nuvlabox - Delete location possible